chore: refactor folders structure
