+++ src/components/Pages/Home/index.js
@@ -1,10 +1,12 @@
 import React from "react";
+import { Slider } from "../../Common/Slider"
+
 
 function Home(){
 
     return (
         <React.Fragment>
-
+            <Slider/>
         </React.Fragment>
     )
 }
